Remove a variable that holds length of array
I thought that it is faster to calculate an array length only once and
then use the variable which holds this length, than to calculate the
array length each loop iteration..
But I've done some tests and found out that JS engine is optimized well
enough to use "i < array.length" in a loop. It's time performance is as
good as the variant with an additional variable before the loop.